+++ R/modebar.R
@@ -6,7 +6,7 @@
 #' @importFrom purrr pmap
 #' @importFrom stringr str_c str_extract_all str_replace_all str_squish str_wrap
 #' @importFrom tidyr drop_na
-roboplotr_modebar <- function(p, title, subtitle, caption, height, width, dateformat, info_text = NULL) {
+roboplotr_modebar <- function(p, title, subtitle, caption, height, width, dateformat, info_text = NULL, modebar) {
 
   if(is.null(title)) {
     dl_title <- "img"
@@ -21,6 +21,7 @@ roboplotr_modebar <- function(p, title, subtitle, caption, height, width, datefo
     plot_title <- list(title$title,subtitle,tf$bold)
     ht <- ifelse(is.null(height), 'null', as.character(height))
     wt <- ifelse(is.null(width), 'null', as.character(width))
+    pie <-ifelse(all(p$trace_types == "pie"),"true","false")
     str_c('
           function(gd, params) {
           let oldlayout = JSON.parse(JSON.stringify(gd.layout))
@@ -35,8 +36,11 @@ roboplotr_modebar <- function(p, title, subtitle, caption, height, width, datefo
           let roboplot_logo = new Image();
           roboplot_logo.src = gd.layout.images[0].source;
           roboplot_logo = roboplot_logo.width / roboplot_logo.height
-          setVerticalLayout({"width": true}, gd, ',layout$main,', ["',plot_title[[1]],'","',plot_title[[2]],'",',tolower(plot_title[[3]]),'], pie_plot = ',if(all(p$trace_types == "pie")) { "true" } else { "false" },', logo = roboplot_logo)
+          $(gd).find("div.modebar").css("display","none")
+          setVerticalLayout({"width": true}, gd, ',layout$main,', ["',plot_title[[1]],'","',plot_title[[2]],'",',tolower(plot_title[[3]]),'], pie_plot = ',pie,', logo = roboplot_logo)
+          setYPositions({"width": true}, gd, ',pie,');
           Plotly.downloadImage(gd, {scale: "1", format: "',layout$type,'", width: ',layout$x,', height: ',layout$y,', filename: "',ttl,layout$suffix,'"});
+          $(gd).find("div.modebar").css("display","initial")
           Plotly.relayout(gd, oldlayout)
           delete oldlayout
           }
@@ -203,10 +207,16 @@ roboplotr_modebar <- function(p, title, subtitle, caption, height, width, datefo
 
   }
 
-  p |> config(
+  p <- p |> config(
     displaylogo = FALSE,
     modeBarButtons = list(unname(btn_list))
   )
+  
+  if(modebar == "constant") {
+    p <- config(p, displayModeBar = T)
+  }
+  
+  p
 }
 
 #' @importFrom DT JS