+++ inst/raw/relayout.js
@@ -1,23 +1,59 @@
-function getVerticalLayout(el, legend_fontsize, height = false, keys, pie_chart, logo = undefined, showfinal = false) {
+function rangeSliderShowHide(el, show = true) {
+  if ('rangeslider' in el.layout.xaxis) {
+      if(el.layout.xaxis.rangeslider.visible != show) {
+        Plotly.relayout(el, {"showlegend" : show, "xaxis.rangeslider.visible": show});
+       } else {
+          Plotly.relayout(el, {"showlegend" : show});
+    }
+  } else {
+        Plotly.relayout(el, {"showlegend" : show});
+  }
 
+}
+
+function rangeSliderHeight(el) {
+    let elslider = 0;
+  if ('rangeslider' in el.layout.xaxis) {
+    if(el.layout.xaxis.rangeslider.visible == true) {
+      elslider_container = $(el).find('.rangeslider-bg')[0].getBBox().height
+      elslider = elslider_container+(el.layout.margin.t*0.1)+(el.layout.margin.b*0.1)
+      let xticks = $(el).find('g.xaxislayer-above')
+      if(xticks.length > 0) {
+        xticks = $(el).find('g.xaxislayer-above')[0].getBBox().height
+        elslider = elslider + (xticks / 10)
+        
+      }
+    }
+  }
+  
+  return (elslider)
+}
+
+function findModeBarHeight(el) {
+    let modebar_ht = 0;
+    let elmodebar = $(el).find('div.modebar');
+    if(elmodebar.length > 0) {
+        if(elmodebar.css("display") != "none") {
+            modebar_ht = elmodebar[0].clientHeight;   
+    }
+  }
+  return modebar_ht
+}
+
+function getVerticalLayout(el, legend_fontsize, height = false, keys, pie_chart, logo = undefined, tidy_legend = false) {
+//  console.log("NEW RELAYOUT")
+//  console.log("margin b init: " + el.layout.margin.b)
+  adjustLegendItems(el, tidy_legend);
   let elcontainer = $(el).find("svg.main-svg")[0].height.animVal.value;
   let eltitle = $(el).find('g.g-gtitle')[0].getBBox();
   let is_yaxis2 = $(el).find('g.xy2').length > 0 ? true: false
   let is_yaxistitle2 = $(el).find('g.g-y2title').length > 0 ? true: false
-  let elslider = 0;
-  if ('rangeslider' in el.layout.xaxis) {
-    if(el.layout.xaxis.rangeslider.visible == true) {
-      elslider = $(el).find('g.infolayer > g.rangeslider-container')[0].getBBox().height
-    }
-  }
+  let elslider = rangeSliderHeight(el)
   let elxticks_default = pie_chart ? 5 : 0
   let elxticks = $(el).find('g.xaxislayer-above')
   if(elxticks.length > 0) { elxticks = elxticks[0].getBBox().height } else { elxticks =  elxticks_default };
-  let modebar_ht = 0;
-  let elmodebar = $(el).find('div.modebar');
-  if(elmodebar.length > 0) {
-    modebar_ht = elmodebar[0].clientHeight;
-  }
+  //console.log("xticks: "+ elxticks)
+  let modebar_ht = findModeBarHeight(el)
   let margin_right = 5
   if(is_yaxis2) {
     let yaxis2 = $(el).find('g.overaxes-above > g.xy2-y')[0].getBBox();
@@ -34,69 +70,51 @@ function getVerticalLayout(el, legend_fontsize, height = false, keys, pie_chart,
   }
   let ellegend = {width: 0, height: 0}
   if ($(el).find('g.legend')[0] != undefined) {
-    ellegend.height =  Math.ceil($(el).find('g.legend')[0].getBBox().height)
-    ellegend.width =  $(el).find('g.legend')[0].getBBox().width
+    let legendbox = $(el).find('g.legend')[0].getBBox()
+    ellegend.height =  Math.ceil(legendbox.height)
+    ellegend.width =  legendbox.width
   };
-  let margin_bottom = ellegend.height + 10 + (elcaption + elxticks + elxtitle + elslider*1);
-  // pie charts do not give correct height by measuring the plot, investigate
-  let elplot = pie_chart ? $(el).find('.pielayer') : $(el).find('.gridlayer');
+  let margin_bottom = ellegend.height + 15 + (elcaption + elxticks + elxtitle);
+  let elplot = pie_chart ? $(el).find('.pielayer') : $(el).find('.nsewdrag');
   if (elplot.length > 0) {elplot = elplot[0].getBBox()};
+ // console.log("margin b: " + margin_bottom)
   if (pie_chart) { elplot.height = elcontainer-margin_bottom-margin_top }
   if (elplot.height == 0) { elplot.height = 1}
   let elimages = $(el).find('g.layer-above > g.imagelayer > image')[0].getBBox();
-  margin_bottom = margin_bottom + logoSpace(logo, elimages, margin_bottom, elxtitle, elxticks, ellegend);
+  let logospace = logoSpace(logo, elimages, margin_bottom, elxtitle, elxticks, ellegend);
+  margin_bottom = margin_bottom + logospace;
+//    console.log("margin b then: " + el.layout.margin.b)
   let images_sizey = (elcontainer * 0.05) / elplot.height;
   el.layout.images[0].sizey = images_sizey
-  let legend_y = -((elxticks + 5 + (elslider*2) + elxtitle) / elplot.height)//((margin_bottom - elcaption + (elslider*2)) / elplot);
-
+//  console.log(el.layout.legend)
+  let legend_y = -((elxticks + 10 + (elslider) + elxtitle) / elplot.height)//((margin_bottom - elcaption + (elslider*2)) / elplot);
   if (legend_y < -2 || (elplot.height < (elcontainer / 4))) {
 //    console.log("TOO TOIT")
-    if ('rangeslider' in el.layout.xaxis) {
-      if(el.layout.xaxis.rangeslider.visible == true) {
-        Plotly.relayout(el, {"showlegend" : false, 'xaxis.rangeslider.visible': false })
-      } else { Plotly.relayout(el, {"showlegend" : false}) }
-    } else {
-      Plotly.relayout(el, {"showlegend" : false})
-    }
+    rangeSliderShowHide(el, false)
     elslider = 0;
     elxticks = $(el).find('g.xaxislayer-above')
     if(elxticks.length > 0) { elxticks = elxticks[0].getBBox().height } else { elxticks = elxticks_default };
-/*    if ($(el).find('g.legend')[0] != undefined) {
-      ellegend.height =  Math.ceil($(el).find('g.legend')[0].getBBox().height + 5)
-      ellegend.width =  $(el).find('g.legend')[0].getBBox().width
-    };*/
-    margin_bottom = 10 + elcaption + elxticks + elxtitle;
-    margin_bottom = margin_bottom + logoSpace(logo, elimages, margin_bottom, elxtitle, elxticks, ellegend);
-    let elplot = pie_chart ? $(el).find('.pielayer') : $(el).find('.gridlayer');
-    if (elplot.length > 0) {elplot = elplot[0].getBBox().height};
+    margin_bottom = 15 + elcaption + elxticks + elxtitle;
+//    margin_bottom = margin_bottom + logoSpace(logo, elimages, margin_bottom, elxtitle, elxticks, ellegend);
+  //    console.log("margin b toiten: " + el.layout.margin.b)
+    let elplot = pie_chart ? $(el).find('.pielayer') : $(el).find('.nsewdrag');
+    if (elplot.length > 0) {
+      elplot = elplot[0].getBBox()
+
+    };
+    if (pie_chart) { elplot.height = elcontainer-margin_bottom-margin_top }
     if (elplot.height == 0) { elplot.height = 1}
     images_sizey = (elcontainer * 0.05) / elplot.height;
-    legend_y = -((margin_bottom - elcaption) / elplot.height);
+    el.layout.images[0].sizey = images_sizey
+    legend_y = -((elxticks + 10 + (elslider) + elxtitle) / elplot.height)//((margin_bottom - elcaption + (elslider*2)) / elplot);
+
   }  else if ((elplot.height > (elcontainer / 1.5)) & el.layout.showlegend == false) {
 //    console.log("NUFF SPACE")
-    if ('rangeslider' in el.layout.xaxis) {
-      if(el.layout.xaxis.rangeslider.visible == false) {
-        Plotly.relayout(el, {"showlegend" : true, "xaxis.rangeslider.visible": true});
-       } else {
-          Plotly.relayout(el, {"showlegend" : true});
-    }
-  } else {
-        Plotly.relayout(el, {"showlegend" : true});
-  }
-    let elslider = 0;
-  if ('rangeslider' in el.layout.xaxis) {
-    if(el.layout.xaxis.rangeslider.visible == true) {
-      elslider = $(el).find('g.infolayer > g.rangeslider-container')[0].getBBox().height
-    }
-  }
-  let elxticks_default = pie_chart ? 5 : 0
+    rangeSliderShowHide(el, true)
+    elslider = rangeSliderHeight(el)
   let elxticks = $(el).find('g.xaxislayer-above')
   if(elxticks.length > 0) { elxticks = elxticks[0].getBBox().height } else { elxticks =  elxticks_default };
-  let modebar_ht = 0;
-  let elmodebar = $(el).find('div.modebar');
-  if(elmodebar.length > 0) {
-    modebar_ht = elmodebar[0].clientHeight;
-  }
+  let modebar_ht = findModeBarHeight(el)
   let margin_top = eltitle.height + modebar_ht + 11;
   let elcaption = $(el).find('g.annotation')[0].getBBox().height + 5;
   let elxtitle = $(el).find('g.g-xtitle')
@@ -105,36 +123,29 @@ function getVerticalLayout(el, legend_fontsize, height = false, keys, pie_chart,
   } else {
     elxtitle = 0
   }
-  let ellegend = {width: 0, height: 0}
+  ellegend = {width: 0, height: 0}
   if ($(el).find('g.legend')[0] != undefined) {
     ellegend.height =  Math.ceil($(el).find('g.legend')[0].getBBox().height)
     ellegend.width =  $(el).find('g.legend')[0].getBBox().width
   };
-  margin_bottom = ellegend.height + 10 + (elcaption + elxticks + elxtitle + elslider*1);
-  margin_bottom = margin_bottom + logoSpace(logo, elimages, margin_bottom, elxtitle, elxticks, ellegend);
-  // pie charts do not give correct height by measuring the plot, investigate
-  elplot = pie_chart ? $(el).find('.pielayer') : $(el).find('.gridlayer');
+  margin_bottom = ellegend.height + 15 + (elcaption + elxticks + elxtitle);
+//  margin_bottom = margin_bottom + logoSpace(logo, elimages, margin_bottom, elxtitle, elxticks, ellegend);
+    //console.log("margin b loosen: " + el.layout.margin.b)
+  elplot = pie_chart ? $(el).find('.pielayer') : $(el).find('.nsewdrag');
   if (elplot.length > 0) {elplot = elplot[0].getBBox()};
   if (pie_chart) { elplot.height = elcontainer-margin_bottom-margin_top }
   if (elplot.height == 0) { elplot.height = 1}
   elimages = $(el).find('g.layer-above > g.imagelayer > image')[0].getBBox();
   images_sizey = (elcontainer * 0.05) / elplot.height;
   el.layout.images[0].sizey = images_sizey
-  legend_y = -((elxticks + 5 + (elslider*2) + elxtitle) / elplot.height)//((margin_bottom - elcaption + (elslider*2)) / elplot);
+  legend_y = -((elxticks + 10 + (elslider) + elxtitle) / elplot.height)//((margin_bottom - elcaption + (elslider*2)) / elplot);
   }
-  let title_y = (elcontainer - (21+modebar_ht)) / elcontainer
-  let low_bound_adjust = elxticks == 0 ? -elcaption : 0
-//  low_bound_adjust = ellegend.height == 0 ? low_bound_adjust+(elcaption) : low_bound_adjust
-  elplot = pie_chart ? $(el).find('.pielayer') : $(el).find('.gridlayer');
+  elplot = pie_chart ? $(el).find('.pielayer') : $(el).find('.nsewdrag');
   if (elplot.length > 0) { elplot = elplot[0].getBBox() };
   if (pie_chart) { elplot.height = elcontainer-margin_bottom-margin_top }
   if (elplot.height == 0) { elplot.height = 1}
-  let low_bound = -((margin_bottom+elslider+low_bound_adjust) / elplot.height)//-Math.min((elcontainer - (elplot.height+margin_top)) / elplot.height,(margin_bottom / elplot.height))
-  let annotations_y = low_bound
-  let images_y = low_bound
   let legend_font_size = (ellegend.height > (elplot.height / 2)) ? legend_fontsize.legend - 2 : legend_fontsize.legend;
   legend_font_size = (ellegend.width > $(el).find("svg.main-svg")[0].width.animVal.value) ? legend_font_size - 2 : legend_font_size;
-
   let yaxis_font_size = legend_fontsize.y;
 
   let yaxis_layer = $(el).find('g.yaxislayer-above')
@@ -148,29 +159,17 @@ function getVerticalLayout(el, legend_fontsize, height = false, keys, pie_chart,
       yaxis_font_size = Math.min(Math.floor(el.layout.yaxis.tickfont.size*1.5), legend_fontsize.y)
     }
   }
-  if(showfinal == true) {console.log('legend ht: ' + ellegend.height +
-                                       ' slider ht: ' + elslider +
-                                       ' plot area ht: ' + elplot.height +
-                                       ' plot caption ht: ' + elcaption +
-                                       ' bottom margin: ' + margin_bottom +
-                                       ' container ht: ' + elcontainer,
-                                     ' legend position: ' + legend_y,
-                                     ' caption position: ' + annotations_y,
-  )}
   let thearray = {
-    'title.y': title_y,
     'legend.y': legend_y,
-    'images[0].y': images_y,
     'images[0].sizey': images_sizey,
     'margin.t': margin_top,
     'margin.b': margin_bottom,
     'margin.r': margin_right,
-    'annotations[0].y': annotations_y,
     'legend.font.size': legend_font_size,
     'yaxis.tickfont.size': yaxis_font_size,
     'yaxis2.tickfont.size': yaxis_font_size
   };
-
+  
   let rearray = keys.reduce(function (obj2, key) {
     if (key in thearray) // line can be removed to make it inclusive
     obj2[key] = thearray[key];
@@ -212,21 +211,44 @@ function logoSpace(ratio, image, margin_b, title_h, xtick_h, legend) {
     }
   }
 
-function adjustLegendItems(gd) {
-  let visdata = gd.data.filter(trace => trace.visible === true || !(trace.hasOwnProperty('visible')));
+function adjustLegendItems(gd, tidy = true) {
+
+ let legendItemTexts = $(gd).find('.legend .legendtext')
+
+if(tidy) {
+    let maxWidth = gd.data.reduce((max, trace) => {
+    let thisLegendItem = legendItemTexts.filter(function() {
+        return $(this).text() === trace.name;
+    });
+    if (thisLegendItem.length > 0) {
+        let thisLegendWidth = thisLegendItem[0].getBBox().width;
+        return Math.max(max, thisLegendWidth);
+    }
+    return max;
+}, 0);
+
+  maxWidth = Math.round(maxWidth * 1.02)
+  gd.layout.legend.entrywidth = maxWidth;
+} else {
   let legendItemTexts = $(gd).find('.legend .legendtext')
-  visdata.forEach(trace => {
+
+  gd.data.forEach(trace => {
     let thisLegendItem = legendItemTexts.filter(function() {
     return $(this).text() === trace.name;
 });
     if(thisLegendItem.length > 0) {
         let thisLegendWidth = thisLegendItem[0].getBBox().width
-        trace.legendwidth = Math.round(thisLegendWidth*1.05)
+        trace.legendwidth = Math.round(thisLegendWidth*1.1)
     }
   })
 }
 
-function setVerticalLayout(eventdata, gd, legend_fontsize, plot_title, pie_chart, logo = undefined) {
+
+}
+
+
+
+function setVerticalLayout(eventdata, gd, legend_fontsize, plot_title, pie_chart, logo = undefined, tidy_legend = false) {
   if ('width' in eventdata | 'autosize' in eventdata) {
     if ('rangeslider' in gd.layout.xaxis) {
       if (gd.layout.xaxis.rangeslider.visible == false) {
@@ -241,9 +263,9 @@ function setVerticalLayout(eventdata, gd, legend_fontsize, plot_title, pie_chart
     let caption_text = gd.layout.annotations[0].text.replace(new RegExp("<br class = 'roboplotr-breaker'>", 'g'), " ")
     Plotly.relayout(gd, {'title.text': title_text, 'annotations[0].text': caption_text})
     let gdtitle = $(gd).find('g.g-gtitle')[0].getBBox().width;
-    let titlespace = pie_chart ? $(gd).find('g.layer-above') : $(gd).find('.cartesianlayer > .xy > .gridlayer');
+    let titlespace = pie_chart ? $(gd).find('g.layer-above') : $(gd).find('.nsewdrag');
     if (titlespace.length > 0) {titlespace = titlespace[0].getBBox().width};
-    let relayout_array = getVerticalLayout(gd, legend_fontsize, false, keys = ['legend.font.size','margin.t','margin.b','legend.y','yaxis.tickfont.size','yaxis2.tickfont.size'], pie_chart = pie_chart, logo = logo)
+    let relayout_array = getVerticalLayout(gd, legend_fontsize, false, keys = ['legend.font.size','margin.t','margin.b','legend.y','yaxis.tickfont.size','yaxis2.tickfont.size'], pie_chart = pie_chart, logo = logo, tidy_legend = tidy_legend)
     if(titlespace <= gdtitle) {
       title_text = "<span>" +
         (plot_title[2] ? "<b>" : "" ) +
@@ -252,19 +274,47 @@ function setVerticalLayout(eventdata, gd, legend_fontsize, plot_title, pie_chart
         "<br><span style='font-size: 75%'>" + plot_title[1] + "</span></span>"
         relayout_array["title.text"] = title_text;
     }
-    adjustLegendItems(gd)
+    rangeSliderShowHide(gd, true);
     Plotly.relayout(gd, relayout_array);
     let logo_width = calculateDisplayedImageSize(logo, $(gd).find('g.layer-above > g.imagelayer > image')[0].getBBox()).width
-    relayout_array = getVerticalLayout(gd, legend_fontsize, false, keys = ['legend.font.size','margin.t','margin.b','margin.r','yaxis.tickfont.size','yaxis2.tickfont.size','images[0].sizey'], pie_chart = pie_chart, logo = logo)
+    relayout_array = getVerticalLayout(gd, legend_fontsize, false, keys = ['legend.font.size','legend.y','margin.t','margin.b','margin.r','yaxis.tickfont.size','yaxis2.tickfont.size','images[0].sizey'], pie_chart = pie_chart, logo = logo, tidy_legend = tidy_legend)
     relayout_array = findCaptionSpace(gd, logo, pie_chart, relayout_array, titlespace);
     Plotly.relayout(gd, relayout_array);
-    relayout_array = getVerticalLayout(gd, legend_fontsize, false, keys = ['legend.font.size','margin.t','margin.b','yaxis.tickfont.size','yaxis2.tickfont.size','images[0].sizey'], pie_chart = pie_chart, logo = logo)
+    relayout_array = getVerticalLayout(gd, legend_fontsize, false, keys = ['legend.font.size','legend.y','margin.t','margin.b','margin.r','yaxis.tickfont.size','yaxis2.tickfont.size','images[0].sizey'], pie_chart = pie_chart, logo = logo, tidy_legend = tidy_legend)
     relayout_array = findCaptionSpace(gd, logo, pie_chart, relayout_array, titlespace);
-    adjustLegendItems(gd)
     Plotly.relayout(gd, relayout_array);
-    Plotly.relayout(gd, getVerticalLayout(gd, legend_fontsize, false, keys = ['images[0].y', 'annotations[0].y', 'margin.t', 'margin.b','legend.y','images[0].sizey','yaxis.tickfont.size','yaxis2.tickfont.size'], pie_chart = pie_chart, logo = logo));
-    Plotly.relayout(gd, getVerticalLayout(gd, legend_fontsize, false, keys = ['images[0].y', 'annotations[0].y', 'legend.y', "margin.b", 'margin.r', 'title.y','yaxis.tickfont.size','yaxis2.tickfont.size'], pie_chart = pie_chart, logo = logo));
+    relayout_array = getVerticalLayout(gd, legend_fontsize, false, keys = ['legend.font.size', 'margin.t', 'margin.b','legend.y','images[0].sizey','yaxis.tickfont.size','yaxis2.tickfont.size'], pie_chart = pie_chart, logo = logo, tidy_legend = tidy_legend);
+    relayout_array = findCaptionSpace(gd, logo, pie_chart, relayout_array, titlespace);
+    Plotly.relayout(gd, relayout_array);
+    relayout_array = getVerticalLayout(gd, legend_fontsize, false, keys = ['legend.font.size', 'margin.t', 'margin.b','legend.y','images[0].sizey','yaxis.tickfont.size','yaxis2.tickfont.size'], pie_chart = pie_chart, logo = logo, tidy_legend = tidy_legend);
+    Plotly.relayout(gd, relayout_array);
+
   }
+}
+
+function setYPositions(eventdata, gd, pie_chart = false) {
+
+    if ('width' in eventdata | 'autosize' in eventdata) {
+      
+  let container = $(gd).find("svg.main-svg")[0].height.animVal.value;
+  let modebar_ht = findModeBarHeight(gd)
+  let title_y = (container - (21+modebar_ht)) / container
+  let plot = pie_chart ? $(gd).find('.pielayer') : $(gd).find('.nsewdrag');
+  if (plot.length > 0) {plot = plot[0].getBBox()};
+ // console.log("margin b: " + margin_bottom)
+  let margin_bottom = gd.layout.margin.b
+  let margin_top = gd.layout.margin.t
+  if (pie_chart) { plot.height = container-margin_bottom-margin_top }
+  let mb = container-plot.height-gd.layout.margin.t
+  let low_bound = -(mb / plot.height)//-Math.min((elcontainer - (elplot.height+margin_top)) / elplot.height,(margin_bottom / elplot.height))
+  let annotations_y = low_bound
+  let images_y = low_bound
+
+//  console.log("low bound: " + low_bound)
+  Plotly.relayout(gd, {"images[0].y": low_bound, "annotations[0].y": low_bound, "title.y": title_y})
+
+    }
+
 }
 
 function findCaptionSpace(gd, logo, pie_chart, relayout_array, titlespace) {
@@ -274,7 +324,7 @@ function findCaptionSpace(gd, logo, pie_chart, relayout_array, titlespace) {
   if(pie_chart) {
       captionspace = titlespace
     } else {
-      let plot_width = $(gd).find('.gridlayer')[0].getBBox().width
+      let plot_width = $(gd).find('.nsewdrag')[0].getBBox().width
       let logo_width = calculateDisplayedImageSize(logo, $(gd).find('g.layer-above > g.imagelayer > image')[0].getBBox()).width
       captionspace = Math.max(plot_width - logo_width, Math.round(logo_width / 2))
     }
@@ -298,6 +348,9 @@ function editShapes(gd, line_label) {
   if(sareashape !== null) {
     gd.layout.shapes[sareashape].y0 = gd.layout.yaxis.range[0]
     gd.layout.shapes[sareashape].y1 = gd.layout.yaxis.range[1]
+    if(gd.layout.shapes[sareashape].xnull == true) {
+      gd.layout.shapes[sareashape].x1 = gd.layout.xaxis.range[1]
+    }
   }
   let zlineshape = findShapeById(gd, 'zeroline');
   if (zlineshape !== null) {
@@ -352,8 +405,8 @@ function stringDivider(str, width, spaceReplacer) {
 }
 
 function yrangeRelayout(eventdata, gd, timerId, trace_sums) {
+
   if (Object.prototype.toString.call(eventdata['xaxis.range']) === '[object Array]') {
-    //console.log(\"rangeslider event!!\");
 		var xRange = gd.layout.xaxis.range;
 		var yRange = gd.layout.yaxis.range;
 		var yInside = [];
@@ -415,14 +468,17 @@ yInside = yInside.concat(yStackInside.map(o => o.val));
 		let yMin = Math.min(...yInside);
 		yMin = yMin < 0 ? yMin * 1.05 : yMin * 0.95;
 		let zline = findShapeById(gd, "zeroline")
+		let sarea = findShapeById(gd, "shadearea")
+		var update = {'yaxis.range': [yMin,yMax]}
 		if(zline != null) {
-		  var update = {
-			'yaxis.range': [yMin,yMax], 'shapes[zline].x0': xRange[0]  // updates the end of the yaxis range
+		  update['shapes['+zline+'].x0'] = xRange[0]
 		}
-		} else {
-		 		var update = {
-			'yaxis.range': [yMin,yMax]//, 'shapes[0].x0': xRange[0]  // updates the end of the yaxis range
-		};
+		if(sarea != null) {
+		  if(gd.layout.shapes[sarea].xnull == true) {
+		    update['shapes['+sarea+'].x1'] = xRange[1]
+		  }
+		  update['shapes['+sarea+'].y0'] = yMin
+		  update['shapes['+sarea+'].y1'] = yMax
 		}
 		Plotly.relayout(gd, update);
 		if (timerId >= 0) {
@@ -438,8 +494,12 @@ yInside = yInside.concat(yStackInside.map(o => o.val));
 	}
 }
 
-function plotlyRelayoutEventFunction(eventdata, gd, legend_fontsize, plot_title, rangeslider_sums, pie_chart, logo) {
+function plotlyRelayoutEventFunction(eventdata, gd, legend_fontsize, plot_title, rangeslider_sums, pie_chart, logo, tidy_legend) {
   timerId = 0;
-  setVerticalLayout(eventdata, gd, legend_fontsize, plot_title, pie_chart, logo);
+//  gd.layout.margin.b = 0
+//  Plotly.redraw(gd)
+  setVerticalLayout(eventdata, gd, legend_fontsize, plot_title, pie_chart, logo, tidy_legend);
+  setYPositions(eventdata, gd, pie_chart);
 	yrangeRelayout(eventdata, gd, timerId, rangeslider_sums);
+//	console.log("margin b: " + gd.layout.margin.b, "; margin t: " + gd.layout.margin.t)
 };